add bulk_insert fn for AANode
